pyproject.toml
[project]
name = "compliance-bdd"
version = "0.1.0"
description = "End-to-end BDD testing for Databricks Unity Catalog functions"
requires-python = ">=3.10"
dependencies = [
    "behave>=1.2.6",
    "databricks-sdk>=0.30.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# Only package src/ — pipelines/ and notebooks/ are workspace artifacts, not packages.
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
